Added method to write/load calibration settings to/from a csv file.
For consistency across applications/trials, calibration settings can be written/read to/from a csv file.

For logging/data-collection purposes, each time the calibration changes, a new line is added to the end of the csv file. This allows us to analyze how a user's calibration settings change over time.

** note that filename paths are specified realtive to the data directory of the project.