A minimum activation threshold setting can now be specified.
In myoelectric control, it is common for a user to experience a small amount of "jitter" caused by the small, involuntary muscle activity that occurs even while the user is attempting to keep their muscles at rest. A common solution to this problem is to implement a minimum activation threshold (i.e., a high-pass filter) that ignores all readings below a specified value.

LibMyoProportional now supports a minimum activation threshold that is:
  - specified by the caller (independently for the left and right calibrations)
  - saved/loaded alongside all other calibration settings
  - applied implicitly to scale all readings before they are returned by poll()

** Note, that the EMG data that is logged using this library's logging features is truly "raw" (i.e., not scaled according to the minimum activation thresholds)